+++ core/initializer/test_execution/test_executor/test_executor.go
@@ -15,6 +15,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis/commons"
 	"github.com/kurtosis-tech/kurtosis/commons/docker_manager"
 	"github.com/kurtosis-tech/kurtosis/initializer/banner_printer"
+	"github.com/kurtosis-tech/kurtosis/initializer/test_execution/output"
 	"github.com/kurtosis-tech/kurtosis/initializer/test_suite_launcher"
 	"github.com/kurtosis-tech/kurtosis/initializer/test_suite_metadata_acquirer"
 	"github.com/palantir/stacktrace"
@@ -34,29 +35,27 @@ No logging to the system-level logger is allowed in this file!!! Everything shou
  */
 
 const (
-	testsuiteContainerDescription = "Testsuite Container"
-
 	networkNameTimestampFormat = "2006-01-02T15.04.05" // Go timestamp formatting is absolutely absurd...
+
+	// When the user presses Ctrl-C on the initializer, we need to forward that on to the API container and tell
+	// it to stop. When that happens, this is how long we'll give the API container to gracefully exit before we
+	// hard-kill the container.
+	apiContainerStopTimeoutAfterParentCtxCancellation = 10 * time.Second
+
+	printTestsuiteLogSectionAsError = false
+
+	// For debugging, we'll sometimes want to print logs from the initializer during the section labelled "testsuite logs"
+	// To distinguish initializer logs from testsuite logs, we add this prefix to loglines that come from the initializer
+	//  rather than the testsuite
+	initializerLogPrefix = "[INITIALIZER] "
 )
-
-/*
-Because a test is run in its own goroutine to allow us to time it out, we need to pass the results back
-	via a channel. This struct is what's passed over the channel.
- */
-type testResult struct {
-	// Whether the test passed or not (undefined if an error occurred that prevented us from retrieving test results)
-	testPassed   bool
-
-	// If not nil, the error that prevented us from retrieving the test result
-	executionErr error
-}
-
 
 /*
 Runs a single test with the given name
 
 Args:
 	executionInstanceId: The UUID representing an execution of the user's test suite, to which this test execution belongs
+	testSetupExecutionCtx: The context in which test setup & execution runs, which can potentially be cancelled on a Ctrl-C
 	ctx: The Context that the test execution is happening in
 	log: the logger to which all logging events during test execution will be sent
 	dockerClient: The Docker client to use to manipulate the Docker engine
@@ -72,7 +71,7 @@ Returns:
 */
 func RunTest(
 		executionInstanceId uuid.UUID,
-		testSetupContext context.Context,
+		testSetupExecutionCtx context.Context,
 		log *logrus.Logger,
 		dockerClient *client.Client,
 		subnetMask string,
@@ -112,7 +111,7 @@ func RunTest(
 		time.Now().Format(networkNameTimestampFormat),
 		executionInstanceId.String(),
 		testName)
-	networkId, err := dockerManager.CreateNetwork(testSetupContext, networkName, subnetMask, gatewayIp)
+	networkId, err := dockerManager.CreateNetwork(testSetupExecutionCtx, networkName, subnetMask, gatewayIp)
 	if err != nil {
 		// TODO If the user Ctrl-C's while the CreateNetwork call is ongoing then the CreateNetwork will error saying
 		//  that the Context was cancelled as expected, but *the Docker engine will still create the networks!!! We'll
@@ -138,7 +137,7 @@ func RunTest(
 	defer freeIpAddrTracker.ReleaseIpAddr(testRunningContainerIp)
 
 	testsuiteContainerId, kurtosisApiContainerId, err := testsuiteLauncher.LaunchTestRunningContainers(
-		testSetupContext,
+		testSetupExecutionCtx,
 		log,
 		dockerManager,
 		networkId,
@@ -155,48 +154,126 @@ func RunTest(
 		return false, stacktrace.Propagate(err, "An error occurred launching the testsuite & Kurtosis API containers for executing the test")
 	}
 
-	// The Kurtosis API will be our indication of whether the test suite container stopped within the timeout or not
-	log.Info("Waiting for Kurtosis API container to exit...")
-	kurtosisApiExitCodeInt64, err := dockerManager.WaitForExit(
-		testTeardownContext,
-		kurtosisApiContainerId)
+	// From this point on, all loglines should be from the testsuite
+	banner_printer.PrintSection(log, "Testsuite Logs", printTestsuiteLogSectionAsError)
+	var logStreamer *output.LogStreamer = nil
+	readCloser, err := dockerManager.GetContainerLogs(testTeardownContext, testsuiteContainerId)
 	if err != nil {
-		return false, stacktrace.Propagate(err, "An error occurred waiting for the exit of the Kurtosis API container: %v", err)
+		log.Errorf("An error occurred getting the testsuite container logs: %v", err)
+	} else {
+		newStreamer := output.NewLogStreamer(log)
+		if startStreamingErr := newStreamer.StartStreamingFromDockerLogs(readCloser); err != nil {
+			log.Errorf("The following error occurred when attempting to stream the testsuite logs: %v", startStreamingErr)
+		} else {
+			log.Tracef("%vTestsuite container log streamer started successfully", initializerLogPrefix)
+			logStreamer = newStreamer
+
+			// Catch-all to make sure we don't leave a thread hanging around in case this function exits abnormally
+			defer logStreamer.StopStreaming()
+		}
 	}
-	kurtosisApiExitCode := int(kurtosisApiExitCodeInt64)
+	defer readCloser.Close()
 
-	// At this point, we may be printing the logs of a stopped test suite container, or we may be printing the logs of
-	//  still-running container that's exceeded the hard test timeout. Regardless, we want to print these so the user
-	//  gets more information about what's going on, and the user will learn the exact error below
-	banner_printer.PrintContainerLogsWithBanners(dockerManager, testTeardownContext, testsuiteContainerId, log, testsuiteContainerDescription)
-
-	acceptExitCodeVisitor, found := api_container_exit_codes.ExitCodeErrorVisitorAcceptFuncs[kurtosisApiExitCode]
-	if !found {
-		return false, stacktrace.NewError("The Kurtosis API container exited with an unrecognized " +
-				"exit code '%v' that doesn't have an accept listener; this is a code bug in Kurtosis",
-			kurtosisApiExitCode)
+	log.Tracef("%vWaiting for API container exit...", initializerLogPrefix)
+	kurtosisApiContainerExitError := waitForApiContainerExit(
+			dockerManager,
+			kurtosisApiContainerId,
+			testSetupExecutionCtx,
+			testTeardownContext)
+	log.Tracef("%vAPI container exited; resulting err is: %v", initializerLogPrefix, kurtosisApiContainerExitError)
+	var stopLogStreamerErr error = nil
+	if logStreamer != nil {
+		log.Tracef("%vStopping testsuite container log streamer...", initializerLogPrefix)
+		stopLogStreamerErr = logStreamer.StopStreaming()
+		log.Tracef("%vStopped testsuite container log streamer", initializerLogPrefix)
 	}
-	visitor := testExecutionExitCodeErrorVisitor{}
-	testStatusRetrievalError := acceptExitCodeVisitor(visitor)
-
-	if testStatusRetrievalError != nil {
-		log.Error("An error occurred that prevented retrieval of the test completion status")
-		return false, testStatusRetrievalError
+	// After this point, we can go back to printing initializer logs
+	banner_printer.PrintSection(log, "End Testsuite Logs", printTestsuiteLogSectionAsError)
+	if stopLogStreamerErr != nil {
+		log.Warnf("An error occurred stopping the log streamer: %v", stopLogStreamerErr)
 	}
-	log.Info("The test suite container running the test exited before the hard test timeout")
+	if kurtosisApiContainerExitError != nil {
+		// NOTE: We haven't stopped the testsuite at this point, but that's okay - it'll get torn down
+		return false, stacktrace.Propagate(err, "An error occurred that prevented retrieval of the test completion status")
+	}
+	log.Info("The test suite container exited as expected")
 
-	// If we get here, the test suite container is guaranteed to have stopped so now we get the exit code
+	// If we got here, then the testsuite container exited as the API container expected, meaning the testsuite
+	//  container is stopped, meaning it's okay to WaitForExit using testTeardownContext to get the testsuite
+	//  container's exit code
+	log.Tracef("Waiting for testsuite container to exit...")
 	testSuiteExitCode, err := dockerManager.WaitForExit(
 		testTeardownContext,
 		testsuiteContainerId)
 	if err != nil {
+		log.Tracef("Received an error waiting for the testsuite container to exit: %v", err)
 		return false, stacktrace.Propagate(err, "An error occurred retrieving the test suite container exit code")
 	}
+	log.Tracef("Testsuite container exited with exit code '%v'", testSuiteExitCode)
 	return testSuiteExitCode == 0, nil
 }
 
 
 // =========================== PRIVATE HELPER FUNCTIONS =========================================
+// Waits for the API container to exit, and returns an error (if any) describing the problem with API container exiting
+func waitForApiContainerExit(
+		dockerManager *docker_manager.DockerManager,
+		kurtosisApiContainerId string,
+		testSetupExecutionCtx context.Context,
+		testTeardownCtx context.Context) error {
+	var kurtosisApiExitCodeInt64 int64
+	kurtosisApiExitCodeInt64, err := dockerManager.WaitForExit(testSetupExecutionCtx, kurtosisApiContainerId)
+	if err != nil {
+		if testSetupExecutionCtx.Err() != context.Canceled {
+			return stacktrace.Propagate(err, "An unexpected error occurred waiting for the exit of the Kurtosis API container")
+		}
+
+		// The parent context was cancelled, indicating that the user pressed Ctrl-C. The API container will have no idea
+		//  that this happened, so we need to stop it (which will result in an error code, in the same way as if
+		//  the user had killed the API container) so the code below can work as expected
+		if err := dockerManager.StopContainer(
+			testTeardownCtx,
+			kurtosisApiContainerId,
+			apiContainerStopTimeoutAfterParentCtxCancellation); err != nil {
+			return stacktrace.Propagate(
+				err,
+				"The test execution context was cancelled (likely due to Ctrl-C), which we forwarded to the API container " +
+					"by requesting that it stop so that it would register the termination signal. However, the below error occurred stopping " +
+					"the API container. This should never happen, and indicates a bug in Kurtosis.",
+			)
+		}
+
+		// The API container will be stopped at this point, so grab the exit code (which should be TERM)
+		//  so that all the code below here works as normal
+		kurtosisApiExitCodeInt64, err = dockerManager.WaitForExit(
+			testTeardownCtx,
+			kurtosisApiContainerId)
+		if err != nil {
+			return stacktrace.Propagate(
+				err,
+				"An error occurred waiting for the API container to exit AFTER we'd successfully stopped it; this is very" +
+					"unexpected and likely indicates a bug in Kurtosis",
+			)
+		}
+	}
+
+	// The Kurtosis API will be our indication of whether the test suite container stopped happily or not
+	kurtosisApiExitCode := int(kurtosisApiExitCodeInt64)
+
+	apiExitCodeVisitorAcceptFunc, found := api_container_exit_codes.ExitCodeErrorVisitorAcceptFuncs[kurtosisApiExitCode]
+	if !found {
+		return stacktrace.NewError(
+			"The Kurtosis API container exited with an unrecognized " +
+				"exit code '%v' that doesn't have an accept listener; this is a code bug in Kurtosis",
+			kurtosisApiExitCode,
+		)
+	}
+	visitor := testExecutionExitCodeErrorVisitor{}
+	if err := apiExitCodeVisitorAcceptFunc(visitor); err != nil {
+		return stacktrace.Propagate(err, "The API container exit code indicated an error")
+	}
+	return nil
+}
 
 
 /*
@@ -208,12 +285,12 @@ func removeNetworkDeferredFunc(
 		log *logrus.Logger,
 		dockerManager *docker_manager.DockerManager,
 		networkId string) {
-	log.Infof("Attempting to remove Docker network with id %v...", networkId)
+	log.Debugf("Attempting to remove Docker network with ID %v...", networkId)
 	if err := dockerManager.RemoveNetwork(testTeardownContext, networkId); err != nil {
 		log.Errorf("An error occurred removing Docker network with ID %v:", networkId)
 		log.Error(err.Error())
 		log.Error("NOTE: This means you will need to clean up the Docker network manually!!")
 	} else {
-		log.Infof("Successfully removed Docker network with ID %v", networkId)
+		log.Debugf("Successfully removed Docker network with ID %v", networkId)
 	}
 }