+++ core/initializer/test_execution/test_executor_parallelizer/test_executor_parallelizer.go
@@ -10,17 +10,19 @@ import (
 	"fmt"
 	"github.com/docker/docker/client"
 	"github.com/google/uuid"
+	"github.com/kurtosis-tech/kurtosis/initializer/banner_printer"
+	"github.com/kurtosis-tech/kurtosis/initializer/test_execution/output"
 	"github.com/kurtosis-tech/kurtosis/initializer/test_execution/test_executor"
 	"github.com/kurtosis-tech/kurtosis/initializer/test_suite_launcher"
-	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
-	"io"
-	"io/ioutil"
 	"os"
 	"os/signal"
-	"strings"
 	"sync"
 	"syscall"
+)
+
+const (
+	logErroneousSystemLogsAsError = true
 )
 
 /*
@@ -72,42 +74,54 @@ func RunInParallelAndPrintResults(
 	close(testParamsChan) // We close the channel so that when all params are consumed, the worker threads won't block on waiting for more params
 	logrus.Debug("All test params loaded into work queue")
 
-	outputManager := newParallelTestOutputManager()
+	// This is where erroneous usages of the system-wide logger will be captured so we can warn the user about them
+	// (e.g. using logrus.Info, when testSpecificLogger.Info should have been used)
+	erroneousSystemLogCaptureWriter := output.NewErroneousSystemLogCaptureWriter()
+	outputManager := output.NewParallelTestOutputManager(logrus.StandardLogger().Out, uint(len(allTestParams)), parallelism)
 
 	logrus.Infof("Launching %v tests with parallelism %v...", len(allTestParams), parallelism)
-
 	disableSystemLogAndRunTestThreads(
 		executionId,
 		ctx,
+		erroneousSystemLogCaptureWriter,
 		outputManager,
 		testParamsChan,
 		parallelism,
 		dockerClient,
 		testsuiteLauncher)
-
 	logrus.Info("All tests exited")
 
-	outputManager.printSummary()
-	return outputManager.getAllTestsPassed()
+	allTestsPassed, err := outputManager.PrintSummary()
+	if err != nil {
+		logrus.Errorf("An error occurred printing the test summary: %v", err)
+		return false
+	}
+
+	capturedMessages := erroneousSystemLogCaptureWriter.GetCapturedMessages()
+	logErroneousSystemLogging(capturedMessages)
+
+	return allTestsPassed
 }
 
 
 func disableSystemLogAndRunTestThreads(
 		executionId uuid.UUID,
 		parentContext context.Context,
-		outputManager *ParallelTestOutputManager,
+		erroneousSystemLogWriter *output.ErroneousSystemLogCaptureWriter,
+		outputManager *output.ParallelTestOutputManager,
 		testParamsChan chan ParallelTestParams,
 		parallelism uint,
 		dockerClient *client.Client,
 		testsuiteLauncher *test_suite_launcher.TestsuiteContainerLauncher) {
-		/*
-		    Because each test needs to have its logs written to an independent file to avoid getting logs all mixed up, we need to make
-    sure that all code below this point uses the per-test logger rather than the systemwide logger. However, it's very difficult for
-    a coder to remember to use 'log.Info' when they're used to doing 'logrus.Info'. To enforce this, we capture any systemwide logger usages
-	during this function so we can show them later.
-	*/
-	outputManager.startInterceptingStdLogger()
-	defer outputManager.stopInterceptingStdLogger()
+	// When we're running tests in parallel, each test needs to have its logs written to an independent file to avoid getting logs all mixed up.
+	// We therefore need to make sure that all code beyond this point uses the per-test logger rather than the systemwide logger.
+	// However, it's very difficult for  a coder to remember to use 'log.Info' when they're used to doing 'logrus.Info'.
+	// To enforce this, we capture any systemwide logger usages during this function so we can show them later.
+	standardLoggerOut := logrus.StandardLogger().Out
+	logrus.StandardLogger().SetOutput(erroneousSystemLogWriter)
+	defer func() {
+		logrus.StandardLogger().SetOutput(standardLoggerOut)
+	}()
 
 	var waitGroup sync.WaitGroup
 	for i := uint(0); i < parallelism; i++ {
@@ -133,7 +147,7 @@ func runTestWorkerGoroutine(
 			parentContext context.Context,
 			waitGroup *sync.WaitGroup,
 			testParamsChan chan ParallelTestParams,
-			outputManager *ParallelTestOutputManager,
+			outputManager *output.ParallelTestOutputManager,
 			dockerClient *client.Client,
 			testsuiteLauncher *test_suite_launcher.TestsuiteContainerLauncher) {
 	// IMPORTANT: make sure that we mark a thread as done!
@@ -141,48 +155,50 @@ func runTestWorkerGoroutine(
 
 	for testParams := range testParamsChan {
 		testName := testParams.TestName
-
-		tempFilename := fmt.Sprintf("%v-%v", executionId, testName)
-		writingTempFp, err := ioutil.TempFile("", tempFilename)
-		if err != nil {
-			emptyOutputReader := &strings.Reader{}
-			executionErr := stacktrace.Propagate(err, "An error occurred creating temporary file to contain logs of test %v", testName)
-			outputManager.logTestOutput(testName, executionErr, false, emptyOutputReader)
-			continue
-		}
-		defer writingTempFp.Close()
-
-		// Create a separate logger just for this test that writes to the test execution logfile
-		log := logrus.New()
-		log.SetLevel(logrus.GetLevel())
-		log.SetOutput(writingTempFp)
-		log.SetFormatter(logrus.StandardLogger().Formatter)
-
 		testsuiteDebuggerHostPortBinding := testParams.DebuggerHostPortBinding
-
-		outputManager.logTestLaunch(testName, testsuiteDebuggerHostPortBinding)
+		testLog := outputManager.RegisterTestLaunch(testName, testsuiteDebuggerHostPortBinding)
 		passed, executionErr := test_executor.RunTest(
 			executionId,
 			parentContext,
-			log,
+			testLog,
 			dockerClient,
 			testParams.SubnetMask,
 			testsuiteLauncher,
 			testsuiteDebuggerHostPortBinding,
 			testName,
 			testParams.TestMetadata)
-		writingTempFp.Close() // Close to flush out anything remaining in the buffer
+		outputManager.RegisterTestCompletion(testName, executionErr, passed)
+	}
+}
 
-		// Create a new FP to read the logfile from the start
-		var testOutputReader io.Reader
-		readingTempFp, err := os.Open(writingTempFp.Name())
-		if err != nil {
-			errorMsg := fmt.Sprintf("An error occurred opening the test's logfile for reading; logs for this test are unavailable:\n%s", err)
-			testOutputReader = strings.NewReader(errorMsg)
-		} else {
-			defer readingTempFp.Close()
-			testOutputReader = readingTempFp
-		}
-		outputManager.logTestOutput(testName, executionErr, passed, testOutputReader)
+/*
+Helper function to print a big warning if there was logging to the system-level logging when there should only have been
+ logging to the test-specific logger
+*/
+func logErroneousSystemLogging(capturedErroneousMessages []output.ErroneousSystemLogInfo) {
+	if len(capturedErroneousMessages) == 0 {
+		return
+	}
+
+	banner_printer.PrintBanner(logrus.StandardLogger(), "Erroneous Logs", logErroneousSystemLogsAsError)
+	logrus.Error("There were log messages printed to the system-level logger during parallel test execution!")
+	logrus.Error("Because the system-level logger is shared and the tests run in parallel, the messages cannot be")
+	logrus.Error(" attributed to any specific test. This is:")
+	logrus.Error("   1) A bug in Kurtosis, and a system-level logger call was used when a test-specific logger")
+	logrus.Error("       should have been used (likely)")
+	logrus.Error("   2) Third-party code calling logrus independently, and there's nothing we can do (unlikely, but possible)")
+	logrus.Error("")
+	logrus.Error("The log message(s) attempted, and the stacktrace(s) of origination, are as follows in the order they were logged:")
+	logrus.Error("")
+
+	for i, messageInfo := range capturedErroneousMessages {
+		banner_printer.PrintSection(logrus.StandardLogger(), fmt.Sprintf("Erroneous Message #%d", i+1), logErroneousSystemLogsAsError)
+		logrus.Error("Message:")
+		logrus.StandardLogger().Out.Write(messageInfo.GetMessage())
+		logrus.StandardLogger().Out.Write([]byte("\n")) // The message likely won't come with a newline so we add it
+		logrus.Error("")
+		logrus.Error("Stacktrace:")
+		logrus.StandardLogger().Out.Write(messageInfo.GetStacktrace())
+		logrus.StandardLogger().Out.Write([]byte("\n")) // The stacktrace likely won't end with a newline so we add it
 	}
 }